+++ dirty_cat/__init__.py
@@ -14,16 +14,15 @@ except ModuleNotFoundError:
         "pkg_resources is not available, dependencies versions will not be checked."
     )
 
+from ._datetime_encoder import DatetimeEncoder
+from ._fuzzy_join import fuzzy_join
+from ._gap_encoder import GapEncoder
+from ._minhash_encoder import MinHashEncoder
+from ._similarity_encoder import SimilarityEncoder
+from ._super_vectorizer import SuperVectorizer
+from ._target_encoder import TargetEncoder
 
-from ._datetime_encoder import DatetimeEncoder  # noqa: E402
-from ._gap_encoder import GapEncoder  # noqa: E402
-from ._minhash_encoder import MinHashEncoder  # noqa: E402
-from ._similarity_encoder import SimilarityEncoder  # noqa: E402
-from ._super_vectorizer import SuperVectorizer  # noqa: E402
-from ._target_encoder import TargetEncoder  # noqa: E402
-
-parent_dir = _Path(__file__).parent
-with open(parent_dir / "VERSION.txt") as _fh:
+with open(_Path(__file__).parent / "VERSION.txt") as _fh:
     __version__ = _fh.read().strip()
 
 
@@ -34,4 +33,5 @@ __all__ = [
     "GapEncoder",
     "DatetimeEncoder",
     "SuperVectorizer",
+    "fuzzy_join",
 ]