+++ CHANGES.rst
@@ -1,13 +1,28 @@
 .. currentmodule:: dirty_cat
 
 Release 0.4.0
-=============
+==============
+
+Major changes
+-------------
+
+* New experimental feature: joining tables using :func:`fuzzy_join` by approximate key matching. Matches are based
+  on string similarities and the nearest neighbors matches are found for each category.
+* **datasets.fetching**: contains a new function :func:`fetch_world_bank_indicator` that can be
+    used to download any indicator from the World Bank Open Data platform. It only needs the
+    indicator ID that can be found on the website.
+* Unnecessary API has been made private: everything (files, functions, classes)
+  starting with an underscore shouldn't be imported in your code.
 
 Minor changes
 -------------
+* Removed example `Fitting scalable, non-linear models on data with dirty categories`.
 
-* Unnecessary API has been made private: everything (files, functions, classes)
-  starting with an underscore shouldn't be imported in your code.
+Bug fixes
+---------
+
+* :class:`MinHashEncoder` now considers `None` as missing values, rather
+  than raising an error.
 
 Release 0.3.0
 =============
@@ -55,7 +70,7 @@ Notes
 Release 0.2.2
 =============
 
-Bug-fixes
+Bug fixes
 ---------
 
 * Fixed a bug in the :class:`SuperVectorizer` causing a `FutureWarning`